requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.0
urllib3==2.4.0
Werkzeug==3.1.3
